añadido todo lo necesario para eliminar los mensajes tan solo falta el mensaje de confirmacion

+++ Acme-CnG/src/main/java/services/PrivateMessageService.java
@@ -1,6 +1,7 @@
 
 package services;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
 
@@ -45,6 +46,7 @@ public class PrivateMessageService {
 		PrivateMessage res;
 
 		res = new PrivateMessage();
+		res.setCopy(true);
 
 		return res;
 	}
@@ -54,6 +56,7 @@ public class PrivateMessageService {
 
 		res = new PrivateMessage();
 		res.setRecipient(c);
+		res.setCopy(true);
 
 		return res;
 	}
@@ -66,7 +69,11 @@ public class PrivateMessageService {
 		result = message;
 		result.setRecipient(message.getRecipient());
 		result.setSender(t);
+		result.setText(message.getText());
+		result.setTitle(message.getTitle());
+		result.setAttachments(message.getAttachments());
 		result.setMoment(new Date(System.currentTimeMillis() - 1000));
+		result.setCopy(message.getCopy());
 		this.validator.validate(result, binding);
 
 		return result;
@@ -99,6 +106,7 @@ public class PrivateMessageService {
 		message1.setAttachments(m.getAttachments());
 		message1.setRecipient(m.getRecipient());
 		message1.setSender(m.getSender());
+		message1.setCopy(false);
 
 		final PrivateMessage message2 = new PrivateMessage();
 		message2.setTitle(m.getTitle());
@@ -107,24 +115,103 @@ public class PrivateMessageService {
 		message2.setAttachments(m.getAttachments());
 		message2.setRecipient(m.getRecipient());
 		message2.setSender(m.getSender());
+		message1.setCopy(true);
 
-		this.privateMessageRepository.save(message1);
+		final Collection<PrivateMessage> cr = m.getRecipient().getRecivedMessages();
+		cr.add(message1);
+
+		final Collection<PrivateMessage> cs = c.getSendedMessages();
+		cs.add(message2);
+
 		this.privateMessageRepository.save(message2);
-		//
-		//		final Collection<PrivateMessage> cr = m.getRecipient().getRecivedMessages();
-		//		cr.add(message1);
-		//		final Collection<PrivateMessage> cs = c.getSendedMessages();
-		//		cs.add(message2);
+		this.privateMessageRepository.save(message1);
 
 	}
 
-	public void delete(final PrivateMessage m) {
+	public void deleteReceived2(final PrivateMessage m) {
 		Assert.notNull(m);
+
 		final Actor principal = this.actorService.findByPrincipal();
+		final Collection<PrivateMessage> recibidos = this.myRecivedMessages(principal.getId());
+
+		final Collection<PrivateMessage> todosPrincipal = new ArrayList<PrivateMessage>();
+		todosPrincipal.addAll(recibidos);
+
+		final Collection<PrivateMessage> todosSender = new ArrayList<PrivateMessage>();
+		todosSender.addAll(this.mySendedMessages(m.getSender().getId()));
+
 		Assert.isTrue(principal.getId() == m.getSender().getId() || principal.getId() == m.getRecipient().getId());
-		this.privateMessageRepository.delete(m);
+		if (recibidos.contains(m)) {
+			recibidos.remove(m);
+			principal.setRecivedMessages(recibidos);
+			this.actorService.save(principal);
+
+		}
+
+		if (!todosPrincipal.contains(m) && !todosSender.contains(m))
+			this.privateMessageRepository.delete(m);
+
+	}
+
+	public void deleteSent2(final PrivateMessage m) {
+		Assert.notNull(m);
+
+		final Actor principal = this.actorService.findByPrincipal();
+		final Collection<PrivateMessage> enviados = this.mySendedMessages(principal.getId());
+
+		final Collection<PrivateMessage> todosPrincipal = new ArrayList<PrivateMessage>();
+		todosPrincipal.addAll(enviados);
+
+		final Collection<PrivateMessage> todosRecipient = new ArrayList<PrivateMessage>();
+		todosRecipient.addAll(this.mySendedMessages(m.getRecipient().getId()));
+
+		Assert.isTrue(principal.getId() == m.getSender().getId() || principal.getId() == m.getRecipient().getId());
+
+		if (enviados.contains(m)) {
+			enviados.remove(m);
+			final Collection<PrivateMessage> nuevos = new ArrayList<PrivateMessage>();
+			nuevos.addAll(enviados);
+			principal.setSendedMessages(nuevos);
+			this.actorService.save(principal);
+
+		}
+
+		if (!todosPrincipal.contains(m) && !todosRecipient.contains(m))
+			this.privateMessageRepository.delete(m);
+
+	}
+
+	public void deleteReceived(final PrivateMessage m) {
+		Assert.notNull(m);
+		this.privateMessageRepository.delete(m.getId());
+
+	}
+
+	public void deleteSent(final PrivateMessage m) {
+		Assert.notNull(m);
+		this.privateMessageRepository.delete(m.getId());
+
 	}
 
 	//Other methods ---------------------------------------------------------
+
+	public Collection<PrivateMessage> mySendedMessages(final int actorId) {
+
+		final Collection<PrivateMessage> sm = this.privateMessageRepository.mySendedMessages(actorId);
+
+		return sm;
+	}
+
+	public Collection<PrivateMessage> myRecivedMessages(final int actorId) {
+
+		final Collection<PrivateMessage> sm = this.privateMessageRepository.myRecivedMessages(actorId);
+
+		return sm;
+	}
+
+	public void flush() {
+		this.privateMessageRepository.flush();
+
+	}
 
 }
